Split ticketmaster_stream.rs into modules by concern
The 642-line file mixed raw Ticketmaster JSON types, this service's
normalized response types, date-window math, the Ticketmaster HTTP
client, event normalization/dedup, and the streaming handler itself
all in one namespace.

Split into ticketmaster_stream/{types,dates,client,normalize}.rs,
with mod.rs left holding just the streaming orchestration
(get_concerts_tm_stream). mod.rs re-exports the public types and the
two normalize functions used elsewhere in the crate
(services/playlist_builder.rs), so every existing
`ticketmaster_stream::*` call site needed zero changes — lib.rs and
playlist_builder.rs diffs are both empty. No behavior change: same
warnings (25, same set — two just moved file), same 30 tests
passing, same clippy/fmt result.
